Using shfmt to format all Bash scripts.

+++ scripts/tf2pdf.sh
@@ -1,5 +1,6 @@
-rosrun tf view_frames
-evince frames.pdf 
+rosrun tf2_tools view_frames.py
+evince frames.pdf 2>/dev/null
+if [[ $? -eq 0 ]]; then
+  rm frames.pdf
+fi
 rm frames.gv
-rm frames.pdf
-
